WIP (failed) on showing/hiding the edit trigger
We should be able to hide the "Edit" button when we enter editing mode, and show it again when we leave. I have the hiding part working, but the re-show turns out to be harder - has to do with passing a call-back to jeditable, which so far hasn't worked for me.